address reviews: clean-up some docstrings, separate receptacle file, clean-up
